pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dotday"
version = "0.1.0"
description = "DotDay — a minimal dynamic wallpaper that tracks your day through dots."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pillow",
  "numpy",
]
license = { text = "AGPL-3.0" }
authors = [ { name = "3xpyth0n" } ]
classifiers = [
  "License :: OSI Approved :: GNU Affero General Public License v3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
]

[project.scripts]
dotday = "dotday:main"
